+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineItemAdapter.h
@@ -20,14 +20,14 @@
 @interface OgreOutlineItemAdapter : OgreTextFindBranch 
 {
     OgreOutlineColumn   *_outlineColumn;
-    id              _item;
-    int             _level;
+    id                  _item;
+    NSInteger           _level;
 }
 
 - (id)initWithOutlineColumn:(OgreOutlineColumn*)anOutlineColumn item:(id)item;
 - (OgreOutlineColumn*)outlineColumn;
-- (void)setLevel:(int)level;
-- (int)level;
+- (void)setLevel:(NSInteger)level;
+- (NSInteger)level;
 - (void)expandItemEnclosingItem:(id)item;
 
 @end
