+++ FindPanel/OgreTextFindComponents/OgreTextFindRoot.m
@@ -37,12 +37,12 @@
 - (BOOL)isHighlightable { return YES; }
 
 /* Getting structural detail */
-- (unsigned)numberOfChildrenInSelection:(BOOL)inSelection
+- (NSUInteger)numberOfChildrenInSelection:(BOOL)inSelection
 {
-    return 1; 
+    return 1;
 }
 
-- (id)childAtIndex:(unsigned)index inSelection:(BOOL)inSelection
+- (id)childAtIndex:(NSUInteger)index inSelection:(BOOL)inSelection
 {
     return _component; 
 }
@@ -76,7 +76,7 @@
     return nil;
 }
 
-- (int)index
+- (NSInteger)index
 {
     return 0;
 }
